requirements.txt
psycopg2==2.9.1
PyNaCl==1.4.0
protobuf==3.17.3
types-protobuf==3.17.3
progressbar2==3.42.0
